Add action with README
New action in a working state.

If ansible provides an endpoint to get the supported versions and all,
then this action may be updated from those instead of hardcoding the
versions.

Doing that for the docker-image can be done by grabbing:
https://github.com/ansible/ansible/blob/devel/test/lib/ansible_test/_data/completion/docker.txt

(taking care of replacing devel by the relevant ansible version)